Moving DeviceModelImageOptions to Shared
