+++ src/IoTHub.Portal.Client/Pages/DeviceModels/CreateDeviceModelPage.razor
@@ -5,6 +5,7 @@
 @using IoTHub.Portal.Models
 @using IoTHub.Portal.Models.v10
 @using IoTHub.Portal.Models.v10.LoRaWAN
+@using IoTHub.Portal.Shared.Constants
 @using IoTHub.Portal.Shared.Models
 
 @attribute [Authorize]
