+++ src/IoTHub.Portal.Tests.Unit/Client/Services/LoRaWanDeviceModelsClientServiceTests.cs
@@ -18,6 +18,7 @@ namespace IoTHub.Portal.Tests.Unit.Client.Services
     using NUnit.Framework;
     using RichardSzalay.MockHttp;
     using IoTHub.Portal.Domain.Options;
+    using Shared.Constants;
 
     [TestFixture]
     public class LoRaWanDeviceModelsClientServiceTests : BlazorUnitTest
