+++ src/IoTHub.Portal.Tests.Unit/Server/Controllers/v1.0/LoRaWAN/LoRaWANDeviceModelsControllerTest.cs
@@ -4,6 +4,7 @@
 namespace IoTHub.Portal.Tests.Unit.Server.Controllers.v1._0.LoRaWAN
 {
     using Portal.Server.Controllers.v1._0.LoRaWAN;
+    using Shared.Constants;
 
     [TestFixture]
     public class LoRaWANDeviceModelsControllerTest : BackendUnitTest
